requirements.txt
Flask
requests
aiohttp
pycryptodome
protobuf